+++ frontend/apps/mobile/components/ReservationModal.tsx
@@ -29,163 +29,9 @@ import {
 import { useTranslation } from "react-i18next";
 import { useAuthContext } from "@/hooks/use-auth-context";
 import { AppColors, FontFamilies, FontSizes } from "@/constants/theme";
-
-// Backdrop that appears instantly when opening and disappears instantly when
-// closing starts — no fade animation in either direction.
-function StaticBackdrop({
-  animatedIndex,
-  style,
-  onPress,
-}: {
-  animatedIndex: SharedValue<number>;
-  style: BottomSheetBackdropProps["style"];
-  onPress: () => void;
-}) {
-  const opacity = useSharedValue(animatedIndex.value > -1 ? 1 : 0);
-
-  useAnimatedReaction(
-    () => animatedIndex.value,
-    (current, previous) => {
-      if (previous === null) return;
-      if (current > -1 && previous <= -1) {
-        // Sheet just started opening — appear immediately.
-        opacity.value = 1;
-      } else if (current < previous && current < 0) {
-        // Sheet just started closing — disappear immediately.
-        opacity.value = 0;
-      }
-    }
-  );
-
-  const animatedStyle = useAnimatedStyle(() => ({ opacity: opacity.value }));
-
-  return (
-    <Animated.View
-      style={[style, animatedStyle, { backgroundColor: "rgba(0,0,0,0.4)" }]}
-    >
-      <Pressable style={StyleSheet.absoluteFill} onPress={onPress} />
-    </Animated.View>
-  );
-}
-
-function formatModalTime(dateStr: string): string {
-  return new Date(dateStr).toLocaleTimeString("en-US", {
-    hour: "numeric",
-    minute: "2-digit",
-    hour12: true,
-  });
-}
-
-function ChildAvatar({
-  child,
-  selected,
-  onPress,
-}: {
-  child: Child;
-  selected: boolean;
-  onPress: () => void;
-}) {
-  const initials = child.name
-    .split(" ")
-    .map((p) => p[0])
-    .join("")
-    .toUpperCase()
-    .slice(0, 2);
-
-  return (
-    <TouchableOpacity
-      onPress={onPress}
-      activeOpacity={0.7}
-      className="items-center gap-1.5"
-    >
-      <View
-        className="w-[52px] h-[52px] rounded-full items-center justify-center"
-        style={{
-          backgroundColor: selected
-            ? AppColors.checkboxSelected
-            : AppColors.surfaceGray,
-          borderWidth: selected ? 2.5 : 2,
-          borderColor: selected
-            ? AppColors.checkboxSelected
-            : AppColors.borderLight,
-          borderStyle: selected ? "solid" : "dashed",
-        }}
-      >
-        <Text
-          style={{
-            fontFamily: FontFamilies.bold,
-            fontSize: FontSizes.base,
-            color: selected ? "#fff" : AppColors.mutedText,
-          }}
-        >
-          {initials}
-        </Text>
-      </View>
-      <Text
-        style={{
-          fontFamily: FontFamilies.regular,
-          fontSize: FontSizes.xs,
-          color: AppColors.mutedText,
-        }}
-        numberOfLines={1}
-      >
-        {child.name.split(" ")[0]}
-      </Text>
-    </TouchableOpacity>
-  );
-}
-
-function EventPreviewSection({
-  occurrence,
-  titleOverride,
-}: {
-  occurrence: EventOccurrence;
-  titleOverride?: string;
-}) {
-  const { t: translate } = useTranslation();
-  const timeLabel = translate("occurrence.classTime", {
-    time: formatModalTime(occurrence.start_time),
-  });
-  return (
-    <View className="items-center mb-6">
-      <View
-        className="w-[120px] h-[120px] rounded-2xl overflow-hidden mb-4"
-        style={{ backgroundColor: AppColors.imagePlaceholder }}
-      >
-        {occurrence.event.presigned_url ? (
-          <Image
-            source={{ uri: occurrence.event.presigned_url }}
-            style={{ width: "100%", height: "100%" }}
-            contentFit="cover"
-          />
-        ) : null}
-      </View>
-      <Text
-        className="text-xl text-center mb-1"
-        style={{ fontFamily: FontFamilies.bold, color: AppColors.primaryText }}
-      >
-        {titleOverride ?? occurrence.event.title}
-      </Text>
-      <Text
-        className="text-sm text-center mb-3"
-        style={{ fontFamily: FontFamilies.regular, color: AppColors.mutedText }}
-      >
-        {timeLabel}
-      </Text>
-      {!!occurrence.event.description && (
-        <Text
-          className="text-sm text-center leading-5"
-          style={{
-            fontFamily: FontFamilies.regular,
-            color: AppColors.secondaryText,
-          }}
-        >
-          {occurrence.event.description}
-        </Text>
-      )}
-    </View>
-  );
-}
+import { ChildAvatar } from "./ChildAvatar";
+import { EventPreviewSection } from "./EventPreviewSection";
+import { StaticBackdrop } from "./StaticBackground";
 
 interface ReservationModalProps {
   visible: boolean;
@@ -293,9 +139,7 @@ export function ReservationModal({
         <EventPreviewSection
           occurrence={occurrence}
           titleOverride={
-            step === "done"
-              ? translate("reservation.completed")
-              : undefined
+            step === "done" ? translate("reservation.completed") : undefined
           }
         />
